Extract config import export utilities
