+++ src/lib/utils/config-export.test.ts
@@ -1,0 +1,29 @@
+import { describe, expect, it } from 'vitest';
+import { exportGhosttyConfig } from '@/lib/utils/config-export';
+import { SPECTRE_VERSION } from '@/lib/version';
+
+describe('exportGhosttyConfig', () => {
+  it('renders Ghostty config output without depending on the Zustand store', () => {
+    const output = exportGhosttyConfig(
+      {
+        'font-family': ['JetBrains Mono', 'Symbols Nerd Font'],
+        keybind: ['ctrl+shift+e=text:FOO=bar', 'clear'],
+        palette: ['0xF=ffffff'],
+        'gtk-custom-css': ['"?required.css"', '?optional.css'],
+        'unknown-option': 'raw = value',
+      },
+      'Dracula',
+    );
+
+    expect(output).toContain(`# Generated by Spectre v${SPECTRE_VERSION}`);
+    expect(output).toContain('# Theme: Dracula');
+    expect(output).toContain('font-family = "JetBrains Mono"');
+    expect(output).toContain('font-family = "Symbols Nerd Font"');
+    expect(output).toContain('keybind = ctrl+shift+e=text:FOO=bar');
+    expect(output).toContain('keybind = clear');
+    expect(output).toContain('palette = 15=#ffffff');
+    expect(output).toContain('gtk-custom-css = "?required.css"');
+    expect(output).toContain('gtk-custom-css = ?optional.css');
+    expect(output).toContain('unknown-option = "raw = value"');
+  });
+});
