+++ src/lib/utils/config-export.ts
@@ -1,0 +1,75 @@
+import type { ConfigValues } from "@/lib/schema/types";
+import { getConfigOption, isPathOption } from "@/lib/utils/config-options";
+import { normalizePaletteEntries } from "@/lib/utils/palette";
+import { SPECTRE_VERSION } from "@/lib/version";
+
+function formatValue(key: string, value: unknown): string {
+  const option = getConfigOption(key);
+
+  if (value === null || value === undefined) return "";
+
+  // Convert UI sentinel back to Ghostty's empty-string enum default.
+  if (value === "default" && option?.type === "enum") {
+    return "";
+  }
+
+  if (option?.type === "boolean") {
+    return value ? "true" : "false";
+  }
+
+  if (typeof value === "string") {
+    if (isPathOption(key) && (value.startsWith("?") || value.startsWith('"?'))) {
+      return value;
+    }
+
+    if (value.includes(" ")) {
+      return `"${value}"`;
+    }
+  }
+
+  return String(value);
+}
+
+export function exportGhosttyConfig(config: ConfigValues, appliedTheme?: string | null): string {
+  const lines: string[] = [
+    `# Generated by Spectre v${SPECTRE_VERSION} - Ghostty Config Generator`,
+    "# https://github.com/imrajyavardhan12/spectre-ghostty-config",
+  ];
+
+  if (appliedTheme) {
+    lines.push(`# Theme: ${appliedTheme}`);
+  }
+
+  lines.push("");
+
+  // Group options by category for organized output.
+  const categories = new Map<string, string[]>();
+
+  for (const [key, value] of Object.entries(config)) {
+    const option = getConfigOption(key);
+    const category = option?.category || "other";
+
+    if (!categories.has(category)) {
+      categories.set(category, []);
+    }
+
+    if (Array.isArray(value)) {
+      const items = key === "palette" ? normalizePaletteEntries(value as string[]) : value;
+      for (const item of items) {
+        categories.get(category)!.push(`${key} = ${formatValue(key, item)}`);
+      }
+    } else {
+      categories.get(category)!.push(`${key} = ${formatValue(key, value)}`);
+    }
+  }
+
+  for (const [category, configLines] of categories) {
+    if (configLines.length > 0) {
+      lines.push(`# ${category.charAt(0).toUpperCase() + category.slice(1)}`);
+      lines.push(...configLines);
+      lines.push("");
+    }
+  }
+
+  return lines.join("\n");
+}
